requirements.txt
customtkinter==5.2.0
pygame==2.5.2
Pillow==10.0.1
pystray==0.19.4